+++ gui_package/main_gui_modules/jsonSupport/__init__.py
@@ -2,7 +2,6 @@
 Json Support module for the PPE GUI
 
 Author: Max Chen
-v0.5.1
 """
 
 from .reporting import generate_report