+++ gui_package/main_gui_modules/recordedBags/__init__.py
@@ -1,0 +1,3 @@
+from .record_dispense_bag import DispenseBagRecorder
+
+__all__ = ['DispenseBagRecorder']